+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore(security): upgrade esbuild version to 0.25.0
 - chore(config): add vpc route table route preventative controls in sample config
 - chore(runtime): updated node runtime to node 20
+- chore(securityhub): update sdk v3
 
 ### Configuration Changes